chore: add proxy like CRA
